third_party/blink/web_tests/accessibility/focusable-span.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<p id="p"><span tabindex="0">hello</span>there</p>

<script>
test(() => {
  let axParagraph = accessibilityController.accessibleElementById('p');
  assert_not_equals(axParagraph, null);
  let axSpan = axParagraph.childAtIndex(0);
  assert_not_equals(axSpan, null);
        assert_equals(axSpan.role, "AXRole: AXGenericContainer");
        assert_true(axSpan.isFocusable);
}, 'Ensure that spans with a tabindex get a focusable state.');
</script>
